camera_client/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
cycler==0.11.0
Deprecated==1.2.13
Django==3.2.18
easyocr==1.6.2
flatbuffers==1.12
fonttools==4.38.0
gast==0.4.0
gcloud==0.18.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
h5py==3.8.0
httplib2==0.21.0
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
joblib==1.2.0
jwcrypto==1.4.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
networkx==2.6.3
ninja==1.11.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
plotly==5.13.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
Pyrebase4==4.6.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
shapely==2.0.1
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.13.0
